- Altera campo ValorPagamento de int para decimal nos DTOs e entidade Pagamento
- Adiciona EnderecoConfiguration com restrições de chave, obrigatoriedade e tamanho
- Expande PagamentoConfiguration para definir restrições e precisão decimal
- Aplica configuração de Endereco no AtelieDosPontinhosDbContext
- Cria repositórios assíncronos para Endereco e Pagamento com operações CRUD, busca por destaque e contagem

+++ AtelieDosPontinhos.Application/DTOs/PagamentoDto.cs
@@ -8,7 +8,7 @@ namespace AtelieDosPontinhos.Application.DTOs
     {
         public int Id { get; set; }
         public int FormadePagamento { get; set; }
-        public int ValorPagamento { get; set; }
+        public decimal ValorPagamento { get; set; }
         public DateTime DataPagamento { get; set; }
     }
 
@@ -16,7 +16,7 @@ namespace AtelieDosPontinhos.Application.DTOs
     {
         
         public int FormadePagamento { get; set; }
-        public int ValorPagamento { get; set; }
+        public decimal ValorPagamento { get; set; }
         public DateTime DataPagamento { get; set; }
     }
 
@@ -24,7 +24,7 @@ namespace AtelieDosPontinhos.Application.DTOs
     {
 
         public int FormadePagamento { get; set; }
-        public int ValorPagamento { get; set; }
+        public decimal ValorPagamento { get; set; }
         public DateTime DataPagamento { get; set; }
     } 
 }
